requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai
moviepy
requests
google-cloud-aiplatform
google-cloud-storage
scipy
numpy